Core/Movement: MotionMaster & MovementGenerators cleaning
Port From (https://github.com/TrinityCore/TrinityCore/commit/14c6a4d8238e4f383557359436b5f73cb730d4ea)

+++ Source/Game/AI/CoreAI/PetAI.cs
@@ -430,7 +430,7 @@ namespace Game.AI
                     ClearCharmInfoFlags();
                     me.GetCharmInfo().SetIsReturning(true);
                     me.GetMotionMaster().Clear();
-                    me.GetMotionMaster().MovePoint(me.GetGUID().GetCounter(), x, y, z);
+                    me.GetMotionMaster().MovePoint((uint)me.GetGUID().GetCounter(), x, y, z);
                 }
             }
             else // COMMAND_FOLLOW
